Moved handleBackendErrors to own util component
